requirements.txt
asgiref==3.4.1
Django==3.2.6
django-cors-headers==3.11.0
django-cors-middleware==1.5.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Pillow==8.3.1
PyJWT==2.3.0
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.3.0
